Change build commands for 3rdparty libs to be no-ops in Windows.
Largely a reverse of a previous commit. We previously removed the quotes
from around some interpolated variables, to instead make them regular
variable expansions, but in their context (they expand to build commands
for third-party libraries) we now know that this doesn't work on all
platforms for all versions of CMake.

A longer explanation of the technical issue is in the code comments.

Review: https://reviews.apache.org/r/37019