Transition Windows builds to use glog v0.3.4.
To get Stout tests to compile on Windows, it is necessary to upgrade the
glog dependency to 0.3.4 (which will allow you to compile with MSVC
1800), and then port glog 0.3.4 to support compiling with MSVC 1900
(which Mesos requires). The code comments explain why we can't compile
with 1800 and link simply have Mesos link to it.

Note that we have supplied the MSVC 1900 port to the glog project, under
pull request #43[1].

The longer story follows:

glog v0.3.3 will not build on Windows for a few reasons that are fixed
in the 0.3.4 release. (For example, `algorithms` is not included in one
file the requires it.)

Since 0.3.4 appears to only fix bugs, and because there appear to be no
major changes in (e.g.) the ABI, I believe this is a safe change to
make. Additionally, when we ported it to compile on MSVC 1900, we didn't
want to re-write the changes they made in 0.3.4 to make it run on MSVC
1800, so this was a natural starting point for us.

I choose not to attempt to upgrade the Linux glog dependency, either in
the CMake solution, or in the autotools solution, for 3 reasons.

(1) I don't want to ask the community reconsider its dependency choices
    on the basis that an experimental Windows port requires a more
    recent version.
(2) We are currently rebundling the glog library as a binary, which
    means that a solution will be somewhat involved (i.e., either we
    have to change the patch command we're running on glog, or we have
    to change the binary)
(3) We can easily support both versions in CMake using the very powerful
    macro system, which will allow us download the dependency on Windows
    machines and use the tarball on other systems.

Footnotes:
[1] https://github.com/google/glog/pull/43

Review: https://reviews.apache.org/r/37020