CMake: Fix MESOS-3250, a dynamic load in Stout tests on OS X.
A few third-party libraries (libev, gmock) do not have `make install`
commands available, so below we have to add our own "install" commands.

The reason is: if we do not, we get runtime library load problems on OS
X. In particular, `dydl` will look for these libraries at the prefix we
passed to `configure` (or in `/usr/local` if we did not pass a prefix
in), but since they don't have a `make install` step, they never get
placed in the prefix folder.

Our solution is to:
  (1) make a lib directory inside the Mesos folder for each of the
      libraries that has no install step, and
  (2) copy all such libraries into their respective directories.

(Note that step (1) is not only convenient, but important: make will add
a `lib` to the end of your prefix path when linking, and since the built
libraries end up in a `.libs` folder, it's not enough to simply pass the
build directory into `configure` as a prefix; so if we're going to move
the libraries, we might as well move them to a library folder.)