CMake: Fix MESOS-3395, MESOS-3394, add canonical third-party lib source.
In the current CMake build system, if we choose to download the
third-party dependencies (instead of un-tar'ing the rebundled versions
of those dependencies) we will download them from the GitHub Mesos
mirror.

It would be better to have a canonical source of third-party
dependencies, independent of the Mesos changes.

This commit points at a new canonical source of third-party tarballs.