[2/2]Generate make batch file to build project in windows.
Generate make batch file to build project in windows.

This commit is the product of two reviews; the first one[1], by
haosdent, was lightly rebased by hausdorff to target a later HEAD, and
then augmented to fix the broken Ubuntu 14.04 build that this commit
caused. Other than those things, the review is functionally identical to
the original, and all credit for the work should go to haosdent.

[1] Original review: https://reviews.apache.org/r/37275

Review: https://reviews.apache.org/r/38540