CMake: Add `CMAKE_NOOP` to common definitions file.
This was originally part of this review[1] by haosdent. It was split out
and moved into the root folder by hausdorff, because we will later reuse
it to build other third-party libraries on Windows that are not only in
libprocess, and hence it belongs outside of the commits that touch
libprocess.

But, the code is essentially haosdent's, so the credit should go to him.

[1] Original review: https://reviews.apache.org/r/37273

Review: https://reviews.apache.org/r/38538