[1/2]CMake: Add libevent version, require Windows builds to use it.
Mesos will need to use libevent to build the process library on Windows,
at least in the immediate term, as libev does not have a standalone
build that works on Windows (i.e., Windows users must integrate the code
directly into their application and compile it as part of their
binaries.) Note that our current plan is eventually to transition to
libevent anyway, so this is just a little ahead of the curve.

This commit will add a default version of libevent, which we will
eventually retrieve from the 3rdparty GitHub repository, which for now
is our "official" distribution channel for out-of-band third-party
dependencies (especially on Windows, which has no package manager).

Additionally, this commit will cause Windows to fail if users do not
specifically opt into using libevent support by passng in  in
`-DENABLE_LIBEVENT`, displaying a message that tells users Windows
builds only support libevent, but that support for libevent in general
is less well-exercised, and so they should use caution.

Review: https://reviews.apache.org/r/38549