[2/2]CMake: Add Windows-specific build targets for APR.
APR is required for Mesos to run. Normally we expect that a user has
obtained APR with their favorite package manager, but on Windows, we
cannot really expect that this is the case. So, we have to rope in the
dependency manually in the CMake build system.

This commit will introduce the necessary build targets to obtain,
configure, build, and install APR for Windows builds. It will ignore APR
and assume it's there for Linux builds.

Review: https://reviews.apache.org/r/38552