CMake: Move to CMake-based build system for many 3rd-party dependencies.
This commit will transition libevent, gmock, and http-parser to each use
their own pure CMake-based build system for CMake builds on Windows.
This is good for a lot of reasons, but the primary benefit is that a
homeneous CMake-based build system is much easier to deal with in
general, particularly if users expect to use a toolchain that is not
"normal" for the Mesos project, like nmake or eclipse.

In the case of libevent, the version we currently support on Windows has
a CMake build solution already in place. For http-parser, we had to make
our own build solution, but we consider this acceptable because it is a
trivial amount of code. For gmock, only version 1.7 has a pure CMake
build system in place, which means that it only became an option to make
this transition after haosdent upgraded.

Review: https://reviews.apache.org/r/39202