Windows: Move ::UUID to stout::UUID to avoid namespace collision.
The `Windows.h` head includes a header that defines the UUID struct for
the DCE RPC API. This is dumped into the global namespace, just like our
UUID implementation found in `stout/uuid.hpp`.

Since this causes a compilation error on Windows, we simply move our
UUID into `stout::`. We further add a `using stout::UUID;` in the file
so that we don't have to change every callsite in the Mesos codebase
that uses it.

Review: https://reviews.apache.org/r/39209