Windows: Moved ::UUID to id::UUID to avoid namespace collision.
`Windows.h` indirectly includes a header that defines a UUID struct for
the DCE RPC API. This is dumped into the global namespace, just like our
UUID implementation found in `stout/uuid.hpp`.

Since this causes a compilation error on Windows, we simply move our
UUID into `id::`. We further add a `using id::UUID;` in the file so that
we don't have to change every callsite in the Mesos codebase that uses
it.

Review: https://reviews.apache.org/r/39209