Windows:[1/3] Moved `os::environ` -> `os::raw::environment`.
Windows standard headers define a macro, `environ`, that holds the
environment data for the executing process. Unfortunately, the existance
of this macro makes it impossible to use a function called
`os::environ`.

Our solution to this problem in this commit is to move the function
family at `os::environ`* to `os::raw::environment`. This family exists
in the `os::raw` because they are the "unstructured" variants of
functions with the same names that exist in the `os` namespace. For
example, `os::raw::environment` returns a `char**` which is the
"unstructured" equivalent of `os::environment`, which returns a
`map<string, string>`.

Review: https://reviews.apache.org/r/39262