Changed commit hook linting to ignore empty diffs.
On Windows, if you attempt to commit an empty changeset, the commit hooks will
attempt to lint the entire repository. This is because we pass blank arguments
to the call to `xargs` that kicks off the C++ linter (e.g., in
`support/pre-commit`). In the git bash, the default behavior of blank arguments
is *not* to ignore them, as it is on certain other platforms. Note that the
`-r` flag is provided to avoid this behavior, but it is only available on
some platforms, and hence is inadmissable here.

Hence, our solution is to check if the results of `git diff` are empty, and
only call `mesos-style.py` if not.

Review: https://reviews.apache.org/r/40195